refactor richard
