+++ src/assets/richard/richard.png
