--- src/components/Celebration.tsx
@@ -1,9 +1,0 @@
-function Celebration() {
-  return (
-    <div style={{display: 'flex',  justifyContent:'center', alignItems:'center', height: '100vh'}}>
-      <h1>Happy Birthday!</h1>
-    </div>
-  );
-}
-
-export default Celebration;
