+++ src/components/richard/CelebrationRichard.tsx
@@ -1,0 +1,16 @@
+function CelebrationRichard() {
+  return (
+    <div
+      style={{
+        display: "flex",
+        justifyContent: "center",
+        alignItems: "center",
+        height: "100vh",
+      }}
+    >
+      <h1>Happy Birthday!</h1>
+    </div>
+  );
+}
+
+export default CelebrationRichard;
